[WIP] Use reflection to call static train method.

+++ core/src/main/java/com/adatao/ddf/DDF.java
@@ -269,7 +269,7 @@ public abstract class DDF extends ALoggable //
     return this.getMetaDataHandler().getNumRows();
   }
 
-  public long getNumColumns() {
+  public int getNumColumns() {
     return this.getSchemaHandler().getNumColumns();
   }
